Make state change idiomatic, revert start length
BEFORE, the way in which extra-keyboardal (yep, it's a word now) state
changes is managed is rather non-idiomatic, using two properties in the
`App` struct, namely `.next_state: GameState` and
`.is_change_state: bool`, to control said changes. Semantically,
`.next_state` really should be an `Option<GameState>`, because it is
being used as such by being paired with `.is_change_state`. When there
_is_ a "next state" to transition to, `.is_change_state` is `true`, and
`.next_state` is set to the appropriate next state, and more tellingly
when there _isn't_ a "next state" to transition to, `.is_change_state`
is `false`, and `.next_state` is populated by a dummy state needed to
satisfy the type as there is no appropriate next state with which to
populate it. This setup was largely chosen because of issues with
mutation from within a `match` control flow operator, to which the
chosen response to fighting with the borrow checker was to obviate the
`match` altogether. However, this solution is somewhat hacky and non-
idiomatic, quite the problem when the entire goal of the project is to
learn idiomatic Rust.

Also, the starting snake length is 15, with the ever-present extra
(16th) segment on the end to indicate where the snake last was for
rendering purposes. This was done to make it faster to lose, making it
easier to test the above changes to `.game_state` upon loss.

AFTER, `.next_state` is now an `Option<GameState>`, with the troubles
within the match now replaced by something far more idiomatic:

    if let Some(game_state) = &mut self.next_state

The borrowing issues were solved by using `std::mem::replace` to move
the reference from within the `Some` of `.next_state` to `.game_state`
and replace it with a generic `GameState` is replaced after the end of
the `if` block with a `None`. Still not entirely idiomatic to set the
value of `.next_state` to `None` during every cycle through right after
its value would be used if it is not already a `None`, but it's better.
Further, the compiler can probably figure out that setting a variable
to `None` hundreds of times when its value is already `None` before the
value finally becomes something that is _not_ `None` is unnecessary,
and so it can be said with decent confidence that this behavior is most
probably optimized out.

Also, the starting snake length has been returned to 2, with the extra
segment now therefore bumped down to the 3rd.