Rendering and food placement changes
BEFORE, the dummy segment at the end of the snake position was rendered
as if it were part of the snake with the square it corresponds to being
filled as a snake square, and then being refilled as a background
square, leading to very minor inefficiency which could easily be fixed.

Also, the food placement random choice excluded this dummy segment,
which could lead to "unwinnable" games even when the game should really
be winnable. This also stood to benefit from being fixed.

AFTER, the dummy segment is no longer included in rendering the snake
body, so it is no longer filled twice.

Also, food placement is now possible on the dummy segment, fixing this
former risk of unwinnable games.